% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lsat.example.dt}
\alias{lsat.example.dt}
\title{Landsat surface reflectance for six sample sites in the Arctic}
\format{
A data.table with 5296 rows and 23 variables
}
\source{
Generated using example code provided in lsatTS::lsat_export_ts()
}
\usage{
lsat.example.dt
}
\description{
A dataset containing Landsat surface reflectance measurements and ancillary
data for six sample sites in the Arctic. These data are used for the examples
included with lsatTS.
}
\keyword{datasets}
