% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsat_calc_spec_index.R
\name{lsat_calc_spec_index}
\alias{lsat_calc_spec_index}
\title{Compute Spectral Indices}
\usage{
lsat_calc_spec_index(dt, si)
}
\arguments{
\item{dt}{Data.table containing surface reflectance data.}

\item{si}{Character string specifying the desired spetral index.}
}
\value{
The input data.table with an appended colunm containing the spectral index
}
\description{
This function computes some widely used spetral indices.
These include the Normalized Difference Vegetation Index (NDVI),
the Enhanced Vegetationd Index (EVI), and 10 others (EVI2, kNDVI, MSI, NBR, NIRv, NDII, NDVSI, NDWI, PSRI, SATVI).
Only one index can be computed at a time.
}
\examples{
my.dt <- lsat_spec_index(my.dt, 'ndvi')
}
