% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsat_calc_trend.R
\name{lsat_calc_trend}
\alias{lsat_calc_trend}
\title{Calculate non-parametric vegetation greenness trends}
\usage{
lsat_calc_trend(
  dt,
  si,
  yrs,
  yr.tolerance = 1,
  nyr.min.frac = 0.66,
  sig = 0.1,
  legend.position = c(0.8, 0.2),
  legend.dir = "horizontal"
)
}
\arguments{
\item{dt}{Data.table with columns including site, year, and the vegetation index of interest}

\item{si}{Spectral index (e.g., NDVI) for which to assess trend}

\item{yrs}{A sequence of years (time period) over which to assess trends (e.g., 2000:2020)}

\item{yr.tolerance}{The number of years that a site's first/last years of observations can differ from the start/end of the user-specified time period ('yrs') for a trend to be computed}

\item{nyr.min.frac}{Fraction of years within the time period for which observations must be available if a trend is to be computed}

\item{sig}{A p-value significance cutoff used to categories trends (e.g., 0.10)}
}
\value{
Data.table with summary of temporal trends by site and a histogram showing summarizing relative changes in vegetation greenness
}
\description{
This function computes a temporal trend in annual time series of vegetation greenness for each sampling site
over a user-specificed time period. This is a wrapper for the zyp.yuepilon() function from the zyp package.
This function will iteratively pre-whiten a time series (i.e., remove temporal autocorrelation) and then compute
Mann-Kendall trend tests and Theil-Sen slope indicators.
}
\examples{
# Forthcoming...
}
