% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsat_clean_data.R
\name{lsat_clean_data}
\alias{lsat_clean_data}
\title{Clean Landsat Data}
\usage{
lsat_clean_data(
  dt,
  cloud.max = 80,
  geom.max = 30,
  sza.max = 60,
  filter.cfmask.snow = T,
  filter.cfmask.water = T,
  filter.jrc.water = T
)
}
\arguments{
\item{dt}{Data.table generated by calling lsat_general_prep().}

\item{cloud.max}{Maximum allowable cloud cover in Landsat scene (percentage).}

\item{geom.max}{Maximum allowable geometric uncertainty (meters).}

\item{sza.max}{Maximum allowable solar zenith angle (degrees).}

\item{filter.cfmask.snow}{(TRUE/FALSE) Remove observations with CFmask flag = snow.}

\item{filter.cfmask.water}{(TRUE/FALSE) Remove observations with CFmask flag = water.}

\item{filter.jrc.water}{(TRUE/FALSE) Remove observations that were ever innundated
based on JRC Global Surface Water Dataset.}
}
\value{
A data.table that includes Landsat observations that met the quality control criteria.
}
\description{
This function filters out observations that exhibit:
(1) clouds and optionally water and snow;
(2) imposibly high reflectance (>1.0) and abnormally low reflectance (<0.005);
(3) scene cloud cover above a user-defined threshold;
(4) geometric uncertainty exceeding a user-defined threshold.
More information about Landsat Quality Bands here:
https://landsat.usgs.gov/collectionqualityband
}
\examples{
data(lsat.example.dt)
lsat.dt <- lsat_general_prep(lsat.example.dt)
lsat.dt <- lsat_clean_data(lsat.dt)
lsat.dt
}
