% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsat_download_ts.R
\name{lsat_download_ts}
\alias{lsat_download_ts}
\title{Download Surface Reflectance Time-Series from the whole Landsat record using rgee}
\usage{
lsat_download_ts(site_coords, export_dir)
}
\arguments{
\item{site_coords}{sfc object containing the site point coordinates for which the time-series should be extracted}

\item{export_dir}{folder path for the export objects (on google drive)}
}
\value{
Execution status
}
\description{
This function grabs the surface reflectance time series from the whole Landsat
record using the Google Earth Engine (account required). These data can then be
processed using the remainder of the lsatTS package workflow.
}
\details{
Plese note: unlike the other functions in this package, this function does NOT
return the time-series as an object, instead these are exported as a CSV to a
GEE users' google drive. This export allows for more flexible scheduling and does
not require the RSession to continue to run in the background while the requests
are processed on the Earth Engine.
}
\examples{


}
