% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsat_evaluate_phenological_max.R
\name{lsat_evaluate_phenological_max}
\alias{lsat_evaluate_phenological_max}
\title{Evaluate Estimates of Annual Phenological Maximum}
\usage{
lsat_evaluate_phenological_max(
  dt,
  vi,
  min.frac.of.max = 0.75,
  zscore.thresh = 3,
  min.obs = 6,
  reps = 10,
  outdir = "output/pheno_max_eval/"
)
}
\arguments{
\item{dt}{Data.table output from lsat_fit_phenologic_curves().}

\item{vi}{Character string specifying the vegetation index (e.g., NDVI) to evaluate.}

\item{min.frac.of.max}{Numeric threshold (0-1) that defines the "growing season" as the seasonal window when the
phenological curves indicate the VI is within a specified fraction of the maximum VI. In otherwords, an observation
is considered to be from the "growing season" when the VI is within a user-specified fraction of the curve-fit growing season maximum VI.}

\item{zscore.thresh}{Numeric threshold specifying the Z-score value beyond which individual observations are filtered before computing the maximum VI.}

\item{min.obs}{Minimum number of observations needed for a site x year to be included in the evaluation (Default = 10)}

\item{reps}{Number of times to bootstrap the assessment (Default = 10)}

\item{outdir}{Output directory into which evaluation data and figure will be writen.}
}
\value{
Data.table
}
\description{
Assess how the number of annual Landsat observations impacts estimates of annual maximum vegetation greenness derived from raw observations and
phenologic modeling. The algorithem computes annual maximum vegetation greenness using site x years with a user-specific number of observations and then
compares these with estimates derived when using progressively smaller subsets of observations. This lets you determine the degree to which
annual estimates of maximum vegetation greenness are impacted the the number of available observations.
}
\examples{
# Forthcoming...
}
