% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsat_format_data.R
\name{lsat_format_data}
\alias{lsat_format_data}
\title{Formats Landsat data for analysis}
\usage{
lsat_format_data(dt)
}
\arguments{
\item{dt}{Data.table with Landsat data exported from Google Earth Engine using lsat_export_ts().}
}
\value{
Data.table with formatted and scaled values.
}
\description{
This function takes Landsat data exported from GEE and formats it
for subsequent use. The function parses sample site coordinates and time period of each measurement,
scales band values, and formats column names as needed for subsequent analysis using the
LandsatTS package.
}
\examples{
data(lsat.example.dt)
lsat.dt <- lsat_format_data(lsat.example.dt)
lsat.dt
}
