% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsat_general_prep.R
\name{lsat_general_prep}
\alias{lsat_general_prep}
\title{Prepare Landsat Data}
\usage{
lsat_general_prep(dt)
}
\arguments{
\item{dt}{Data.table containing Landsat data extracted from Google Earth Engine.}
}
\value{
Data.table with formatted and scaled values.
}
\description{
This function parses the (1) Landsat satellite; (2) Landsat collection; (3) year; and (4) day of year from the Landsat ID.
The function also renames and rescales bands 1 - 7, taking into account differences between Landsat 5 / 7 and 8
}
\examples{
parsed.dt <- lsat_general_prep(gee.dt)
}
