% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsat_get_pixel_centers.R
\name{lsat_get_pixel_centers}
\alias{lsat_get_pixel_centers}
\title{Landsat 8 pixels centers within a polygon}
\usage{
lsat_get_pixel_centers(polygon_sf, buffer)
}
\arguments{
\item{polygon_sf}{Simple feature collection (or simple feature) with the polygon's geometry.}

\item{buffer}{Buffer surrounding the geometry to be included. Specified in m.}
}
\value{
sfc of point geometries for Landsat 8 pixel centers within the polygon for use in lsat_download_ts
}
\description{
A convenience helper function that determines the Landsat 8 grid (pixel)
centers within a polygon and a surrounding buffer.
}
\details{
Will fail gracefully for large polygons or polygons extending across multiple
Landsat WRS tiles. Errors/warnings will be issues in this case.
}
\examples{

my_polgyon <-

}
