% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsat_get_pixels_in_buffer.R
\name{lsat_get_pixels_in_buffer}
\alias{lsat_get_pixels_in_buffer}
\title{Landsat 8 pixel centres within a buffer around point coordinates}
\usage{
lsat_get_pixels_in_buffer(
  point_coord_sf,
  buffer_dist = 0,
  sample_id_from = "sample_id",
  n_threads = 1
)
}
\arguments{
\item{point_coord_sf}{simple feature collection of points with a column that
indicates the sample_id (speciefied with option sample_id_from).}

\item{buffer_dist}{buffer distance to be applied to point coordinates. Note:
a distance of 15 m is automatically added. So when buffer distance = 0 the
nearest pixel centre to the coordinate(s) is identified. If buffer distance =
30, then the centre coordiantes of the pixels in the 3 x 3 neigbourhood
that is centred on the coordinate(s) os returned.}

\item{sample_id_from}{column that provides the unique sample identifier.
Defaults to "sample_id". Pixels centers within the buffer are returned by
appending a counter "_x" to this column.}

\item{n_threads}{the number of threads to be used if parallel processing is
desired. Recommended for large datasets.}
}
\value{
Simple feature collection of points that denote pixel centres within
buffer around input coordinates. Unique identifiers are provided by
appending a numerical counter "_x" to original identifiers in column
specified in the sample_id argument.
}
\description{
Helper function that uses lsat_get_pixel_centers() to determine all Landsat 8
pixel centre coordinates within a buffer around a set of point coordinates.
}
\examples{

# Using sf, dplyr and rgee
library(sf)
library(dplyr)
library(rgee)

# Initialize rgee
ee_Initialize()

# Generate test points
test_points_sf <- st_sfc(st_point(c(-149.6026, 68.62574)),
                         st_point(c(-149.6003, 68.62524)),
                         st_point(c(-75.78057, 78.87038)),
                         st_point(c(-75.77098, 78.87256)),
                         st_point(c(-20.56182, 74.47670)),
                         st_point(c(-20.55376, 74.47749)), crs = 4326) \%>\%
  st_sf() \%>\%
  mutate(sample_id = c("toolik_1",
                      "toolik_2",
                      "ellesmere_1",
                      "ellesmere_1",
                      "zackenberg_1",
                      "zackenberg_2"),
         region = c("toolik", "toolik",
                    "ellesmere", "ellesmere",
                    "zackenberg", "zackenberg"))

# Retrieve pixel centres in 3 x 3 neighbourhood using parallel processing
# and 4 threads

}
\author{
Jakob J. Assmann
}
