% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsat_summarize_data_avail.R
\name{lsat_summarize_data_avail}
\alias{lsat_summarize_data_avail}
\title{Summarize Availability of Landsat data by Site}
\usage{
lsat_summarize_data_avail(dt)
}
\arguments{
\item{dt}{Data.table with columns named "site" and "year"}
}
\value{
Data.table summarizing for each site the first, last, and number of years with observations,
the minimum and maximum number of observations in a year, the total number of observations across years.
}
\description{
This little function summarizes the temporal period and availability of observations at each site
}
\examples{
# summary.dt <- lsat_summarize_data_avail(dt)
}
