% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsat_summarize_growing_seasons.R
\name{lsat_summarize_growing_seasons}
\alias{lsat_summarize_growing_seasons}
\title{Summarize a Vegetation Index for each Growing Season}
\usage{
lsat_summarize_growing_seasons(
  dt,
  si,
  min.frac.of.max = 0.75,
  zscore.thresh = 3
)
}
\arguments{
\item{dt}{Data.table generated by the function lsat_fit_phenologic_curves().}

\item{si}{Character string specifying the vegetation index to summarize (e.g., NDVI).}

\item{min.frac.of.max}{Numeric threshold (0-1) that defines the "growing season" as the seasonal window when the
phenological curves indicate the SI is within a specified fraction of the maximum SI. In other words, an observation
is considered to be from the "growing season" when the SI is within a user-specified fraction of the curve-fit growing season maximum SI.}

\item{zscore.thresh}{Numeric threshold specifying the Z-score value beyond which individual observations are filtered before summarizing growing season SI.}
}
\value{
Data.table summarizing annual growing season conditions based on a vegetation index.
}
\description{
This function not only computes mean, median, and 90th percentile of a spectral index (SI)
using observations for a user-specified "growing season," but also estimates the annual maximum VI and associated day of year
using phenology modeling and growing season observations.
}
\examples{
# Forthcoming...
}
