% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsat_calc_trend.R
\name{lsat_calc_trend}
\alias{lsat_calc_trend}
\title{Calculate non-parametric vegetation greenness trends}
\usage{
lsat_calc_trend(dt, vi, yrs, sig = 0.1)
}
\arguments{
\item{dt}{Data.table with columns including site, year, and the vegetation index of interest}

\item{vi}{Vegetation index (e.g., NDVI) for which to assess trend}

\item{yrs}{A sequence of years over which to evaluate a trend (e.g., 2000:2020)}

\item{sig}{A p-value significance cutoff used to categories trends (e.g., 0.10)}
}
\value{
Data.table with summary of temporal trend by site
}
\description{
This function computes a temporal trend in annual time series of vegetation greenness for each sampling site
over a user-specificed time period. This is a wrapper for the zyp.yuepilon() function from the zyp package.
This function will iteratively pre-whiten a time series (i.e., remove temporal autocorrelation) and then compute
Mann-Kendall trend tests and Theil-Sen slope indicators.
}
\examples{
Forthcoming...
}
