% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsat_calibrate_rf.R
\name{lsat_calibrate_rf}
\alias{lsat_calibrate_rf}
\title{Cross-Calibrate Landsat Sensors using Random Forests Models}
\usage{
lsat_calibrate_rf(
  dt,
  band,
  doy.rng,
  min.obs,
  frac.train = 0.75,
  overwrite.col = F,
  outfile.id = band,
  outdir
)
}
\arguments{
\item{dt}{Data.table containing the band or spectral index to cross calibrate.}

\item{band}{Character string matching the column name of the band or spectral index to cross-calibrate.}

\item{doy.rng}{Sequence of numbers specifying the Days of Year (Julian Days) to use for model development.}

\item{min.obs}{Minimum number of paired, seasonally-matched observations from Landsat 7 and Landsat 5/8 required to include a sampling site.}

\item{frac.train}{Fraction of sites to use for training the random forest models. The remaining sites are used for model cross-validation.}

\item{overwrite.col}{(True/False) Overwrite existing column or (by default) append cross-calibrated data as a new column?}

\item{outfile.id}{Identifier used when naming output files. Defaults to the input band, but can be specified if needed such as when performing Monte Carlo simulations.}

\item{outdir}{Output directory (created if necessary) to which multiple files will be written.
The files include the (1) fitted random forest models as an R object, (2) evaluation data in a csv file, (3) summary of
model cross-validation in a csv file, and (4) multi-panel scatter plot comparing sensors pre- and post-calibration in jpeg format.
If cross-calibrating both Landsat 5 and 8, then the function returns files for both sensors.}
}
\value{
The input data.table with an appended column titled band.xcal, where "band" is your specified band or spectral index
}
\description{
There are systematic differences in spectral indices (e.g., NDVI) among Landsat 5, 7, and 8 (Landsat Collection 1).
It is important to address these differences before assessing temporal trends in spectral data.
Failure to address these differences can, for instance, introduce artificial positive trends into NDVI time-series that are
based on measurements from multiple Landsat sensors (Ju and Masek 2016, Roy et al. 2016, Berner et al. 2020).

This function cross-calibrates individual bands or spectral indices from Landsat 5/8 to match Landsat 7.
Landsat 7 is used as a benchmark because it temporally overlaps with the other two sensors.
Cross-calibration can only be performed on one band or spectral index at a time.
The approach involves determining the typical reflectance at a site during a portion of the growing season using
Landsat 7 and Landsat 5/8 data that were collected the same years.
A random forest model is then trained to predict Landsat 7 reflectance from Landsat 5/8 reflectance.
To account for potential seasonal and regional differences between sensors, the random forest models also
include as covariates the midpoint of each 15-day period (day of year) and the spatial coordinates of each sampling site.
This approach can handle non-linear relationships, but is most suitable when working with data from 100s to preferably 1000s of sampling sites.

The specific steps to cross-calibrating sensors include:
(1)	Identify the years when both Landsat 7 and Landsat 5/8 measured surface reflectance at a sampling site.
(2)	Pool the reflectance measurements across those years and compute 15-day moving median reflectance over the course of the growing season for each sensor and sampling site.
(3)	Exclude 15-day periods with fewer than a specified number of measurements from both sets of sensors and then randomly select one remaining 15-day period from each sampling site.
(4)	Split the data into sets for model training and evaluation.
(5)	Train Random Forest models that predict Landsat 7 reflectance based on Landsat 5/8 reflectance. The models also account for potential seasonal and regional differences between sensors by including as covariates the midpoint of each 15-day period (day of year) and the spatial coordinates of each sampling site.
(6)	Fit the random forest models using the ranger package.
See Berner et al. (2020) for a full description of the approach.
}
\examples{
# lsat.dt <- lsat_xcal_rf(lsat.dt, band = 'ndvi', doy.rng = 152:243, min.obs = 5, frac.train = 0.75, outfile.id = 'ndvi', outdir ='data/lsat_site_data/sensor_xcal/ndvi/')
}
