% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsat_clean_data.R
\name{lsat_clean_data}
\alias{lsat_clean_data}
\title{Clean Landsat surface reflectance data}
\usage{
lsat_clean_data(
  dt,
  cloud.max = 80,
  geom.max = 30,
  sza.max = 60,
  filter.cfmask.snow = T,
  filter.cfmask.water = T,
  filter.jrc.water = T
)
}
\arguments{
\item{dt}{Data.table generated by calling lsat_general_prep().}

\item{cloud.max}{Maximum allowable cloud cover in Landsat scene (percentage).}

\item{geom.max}{Maximum allowable geometric uncertainty (meters).}

\item{sza.max}{Maximum allowable solar zenith angle (degrees).}

\item{filter.cfmask.snow}{(TRUE/FALSE) Remove measurements with CFmask flag = snow.}

\item{filter.cfmask.water}{(TRUE/FALSE) Remove measurements with CFmask flag = water.}

\item{filter.jrc.water}{(TRUE/FALSE) Remove sample sites that were ever innundated
based on the maximum surface water extent variable from the
JRC Global Surface Water Dataset.}
}
\value{
A data.table that includes Landsat measurements that met the quality control criteria.
}
\description{
This function enables users to filter out surface reflectance
measurements that exhibit:
(1) clouds, cloud shadows, snow, or water flagged by the CFMask algorithm;
(2) surface water over the Landsat record;
(2) impossibly high reflectance (>1.0) and abnormally low reflectance (<0.005);
(3) scene cloud cover above a user-defined threshold;
(4) geometric uncertainty above a user-defined threshold;
(5) solar zenith angle above a user-defined threshold.
}
\examples{
data(lsat.example.dt)
lsat.dt <- lsat_general_prep(lsat.example.dt)
lsat.dt <- lsat_clean_data(lsat.dt)
lsat.dt
}
