% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsat_fit_phenologic_curves.R
\name{lsat_fit_phenological_curves}
\alias{lsat_fit_phenological_curves}
\title{Characterize Land Surface Phenology using Vegetation Index Time Series}
\usage{
lsat_fit_phenological_curves(
  dt,
  vi,
  window.yrs = 5,
  window.min.obs = 10,
  vi.min = 0,
  spar = 0.7,
  pcnt.dif.thresh = 100,
  spl.fit.outfile = F,
  progress = T
)
}
\arguments{
\item{dt}{Data.table with a multi-year time series a vegetation index}

\item{vi}{Character string specifying the vegetation index (e.g., NDVI) to use for determining surface phenology. This must correspond
to an existing colunm in the data.table.}

\item{window.yrs}{Number specifying the focal window width in years that is used when pooling data to fit cubic splines (use odd numbers).}

\item{window.min.obs}{Minimum number of focal window observations necessary to fit a cubic spline.}

\item{vi.min}{Minimum value of vegetation index necessary for observation to be used when fitting cubic splines}

\item{spar}{Smoothing paramater passed to smooth.spline(), typically around 0.65 - 0.75 for this application.}

\item{pcnt.dif.thresh}{Allowable percent difference (0-100) between individual observations and fitted cubic spline.
Observations that differ by more than this threshold are filtered out and the cubic spline is iteratively re-reft.}

\item{spl.fit.outfile}{(Optional) Name of output csv file containing the fitted cubic splines for each site.
Useful for subsequent visualization}

\item{progress}{(TRUE/FALSE) Print a progress report?}
}
\value{
Data.table that provides, for each observation, information on the phenological conditions for that specific day of year during the focal period.
These data can then be used to estimate annual maximum vegegation index and other growing season metrics using lsat_summarize_growing_season().
}
\description{
This function characterizes seasonal land surface phenology at each sampling site using time series of spectral vegetation indices (e.g., NDVI).
The underlying algorithm was construted to facilitate estimating annual maximum vegetation greenness (vegetation index)
The function returns information about typical phenology at a site and about the timing of an individual observation relative.
Please note that this function was designed for sitations where the seasonal phenology is hump shaped. If you are using a spetral index that is
typcially negative (e.g., Normalized Difference Water Index) then multiply the index by -1 before running this function, then backtransform
your index after running the lsat_summarize_growing_seasons() function.
}
\examples{
# To come...

}
