% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsat_general_prep.R
\name{lsat_general_prep}
\alias{lsat_general_prep}
\title{Prepare Landsat data for analysis}
\usage{
lsat_general_prep(dt)
}
\arguments{
\item{dt}{Data.table with Landsat data exported from Google Earth Engine using lsat_export_ts().}
}
\value{
Data.table with formatted and scaled values.
}
\description{
This function parses sample site coordinates and time period of each measurement,
scales band values, and formats column names as needed for subsequent analysis using the
lsatTS package.
}
\examples{
data(lsat.example.dt)
lsat.dt <- lsat_general_prep(lsat.example.dt)
lsat.dt
}
